Started on setting up framework for application layer.
Most of the work done in this commit is incolmplete and just a starting
place, work in progress.

- Created header for the ADCS application module. This will contain the
adcs tasks and high level functions, like the control loop etc. This is
just an example or template, it is not complete.

- Created the CSP server module. This will receive messages and send
them to the main task, in the proper queue. This task should be
functional currently, but needs testing.

- Created the Operations module file and header. This will contain the
main task and operations state machine. This is currently incomplete.