Successfully tested IAP programming recovery.
- Modifications to allow uploading a golden image and a update image.
- Only the proper firmware is erased when uploading a new image.
- Hard coded the SPI directory needed for programming recovery.

-With this code and the libero project from the next commit, the
programming recovery was tested by uploading a golden image and updated
image. Then iap was initiated and power was removed during this process.
When power was restored, then golden image was programmed and this was
confirmed by using logic analyzer to see traffic between the processor
and flash memory, and by the different software version number printed
by the software.